+++ frontend/src/routes/(root)/(logged)/schedules/+page.svelte
@@ -1,7 +1,6 @@
 <script lang="ts">
 	import {
 		ScheduleService,
-		JobService,
 		type ScheduleWJobs,
 		type WorkspaceDeployUISettings,
 		WorkspaceService
@@ -45,6 +44,7 @@
 	import { onMount } from 'svelte'
 	import DeployWorkspaceDrawer from '$lib/components/DeployWorkspaceDrawer.svelte'
 	import { ALL_DEPLOYABLE, isDeployable } from '$lib/utils_deployable'
+	import { runScheduleNow } from '$lib/components/triggers/scheduled/utils'
 
 	type ScheduleW = ScheduleWJobs & { canWrite: boolean }
 
@@ -103,36 +103,6 @@
 		} catch (err) {
 			sendUserToast(`Cannot ` + (enabled ? 'enable' : 'disable') + ` schedule: ${err.body}`, true)
 			loadSchedules()
-		}
-	}
-
-	async function runScheduleNow(
-		path: string,
-		schedulePath: string,
-		isFlow: boolean
-	): Promise<void> {
-		try {
-			const runByPath = isFlow ? JobService.runFlowByPath : JobService.runScriptByPath
-			const args = (
-				await ScheduleService.getSchedule({
-					workspace: $workspaceStore!,
-					path: schedulePath
-				})
-			).args
-			const run = await runByPath({
-				path,
-				requestBody: args ?? {},
-				workspace: $workspaceStore!
-			})
-
-			sendUserToast(`Schedule ${path} will run now`, false, [
-				{
-					label: 'Go to the run page',
-					callback: () => goto('/run/' + run + '?workspace=' + $workspaceStore)
-				}
-			])
-		} catch (err) {
-			sendUserToast(`Cannot run schedule now: ${err.body}`, true)
 		}
 	}
 
@@ -200,18 +170,18 @@
 							x.path.startsWith(ownerFilter + '/') &&
 							filterItemsPathsBaseOnUserFilters(x, selectedFilterKind, filterUserFolders) &&
 							filterItemsBasedOnEnabledDisabled(x, filterEnabledDisabled)
-				  )
+					)
 				: schedules?.filter(
 						(x) =>
 							x.script_path.startsWith(ownerFilter + '/') &&
 							filterItemsPathsBaseOnUserFilters(x, selectedFilterKind, filterUserFolders) &&
 							filterItemsBasedOnEnabledDisabled(x, filterEnabledDisabled)
-				  )
+					)
 			: schedules?.filter(
 					(x) =>
 						filterItemsPathsBaseOnUserFilters(x, selectedFilterKind, filterUserFolders) &&
 						filterItemsBasedOnEnabledDisabled(x, filterEnabledDisabled)
-			  )
+				)
 
 	$: if ($workspaceStore) {
 		ownerFilter = undefined
@@ -221,10 +191,10 @@
 		selectedFilterKind === 'schedule'
 			? Array.from(
 					new Set(filteredItems?.map((x) => x.path.split('/').slice(0, 2).join('/')) ?? [])
-			  ).sort()
+				).sort()
 			: Array.from(
 					new Set(filteredItems?.map((x) => x.script_path.split('/').slice(0, 2).join('/')) ?? [])
-			  ).sort()
+				).sort()
 
 	$: items = filter !== '' ? filteredItems : preFilteredItems
 
@@ -463,7 +433,7 @@
 																})
 															}
 														}
-												  ]
+													]
 												: []),
 											{
 												displayName: 'View runs',
@@ -483,7 +453,7 @@
 												displayName: 'Run now',
 												icon: Play,
 												action: () => {
-													runScheduleNow(script_path, path, is_flow)
+													runScheduleNow(script_path, path, is_flow, $workspaceStore!)
 												}
 											},
 											{
